% Auto-generated: do not edit by hand
\name{dvcVega}

\alias{dvcVega}

\title{Vega component}

\description{

}

\usage{
dvcVega(id=NULL, opt=NULL, spec=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{opt}{Named list. Vega-Embed options}

\item{spec}{Named list. A Vega or Vega-Lite spec}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

